+++ c/testu01/unif01_stubs.h
@@ -1,0 +1,18 @@
+#include <string.h>
+
+#define CAML_NAME_SPACE
+#include <caml/mlvalues.h>
+#include <caml/memory.h>
+#include <caml/callback.h>
+#include <caml/bigarray.h>
+#include <caml/custom.h>
+
+#include <unif01.h>
+
+static struct custom_operations unif01_Gen_boxed;
+
+#define unif01_Gen_unbox(bgen) (* (unif01_Gen**) Data_custom_val(bgen))
+
+value caml_unif01_CreateExternGenBits(value name, value bits);
+
+value caml_unif01_DeleteExternGenBits(value bgen);
